Use email to identify if our confirmation code is correct, dummy varialbe for email  #23

+++ src/Confirmation.php
@@ -4,22 +4,25 @@
 
   if (isset($_POST['submit'])) {
 	  $confirmation = $_POST['confirmation'];
-	  //write custom function to check the table for our code
-	  // we dont hace access to the table... loginsql doesnt allow us to get any variable out of it or do anything unless we are login.php
-	  
-	  $code = $confirmation;
-	  $time = strtotime(ConfirmationCode::get_time($code));
-	  $newTime = time();
-	  $endTime = strtotime("+15 minutes", $time);
-	  if($endTime>$newTime){
-		//failure
-		echo "<br> You have input an expired confirmation code.<br>";
+	  $conn = LoginSQL::sqlConnect();
+	  $email="lukemcda@buffalo.edu";//dummy variable until we can get actual email
+	  $checkConfirm = LoginSQL::getConfirmCode($conn, $email);
+	  if($checkConfirm == $confirmation){
+		  $code = $confirmation;
+		  $time = strtotime(ConfirmationCode::get_time($code));
+		  $newTime = time();
+		  $endTime = strtotime("+15 minutes", $time);
+		  if($endTime>$newTime){
+			//failure
+			echo "<br> You have input an expired confirmation code.<br>";
+		  }
+		  else{
+			  //success
+			  echo "<br> You have input a correct confirmation code.<br>";
+		  }
 	  }
 	  else{
-		  //success
-		  echo "<br> You have input a correct confirmation code.<br>";
+		 echo "<br> You have input an invalid confirmation code.<br>";
 	  }
-	  
-	  
   }
 ?>
